Comments - Defect Number: 2804; Defect RecordID: 2800; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Opportunities: Create "Discussion Review 2" view; Defect Description: This is an alternative view to the already listed "Discussion Review 1" view. It should be simpler to implement, and it borrows minor functionality from the other view.
In this view, the Opportunity page will remain mostly the same. Combine the names from Salesperson and Owner fields into a "Sales Team" column where the names are stacked and unlink any names there. Remove the Days Old and Last Change columns. In their place, create a "Note" column at a fixed height. The width can dynamically resize as the page is maximized. In the Note box there should be text in grey italics that reads "To add a note, click here and begin typing. Your note will be saved automatically." Whatever the user enters into that box should be saved in PM as a note, but not until the user either clicks into another note box or as soon as it can be in the background (maybe use an autosave every n # of seconds?) - whatever the save method, the cursor should not disappear or change focus. NOTE: It doesn't appear in the image below, but the Est. Revenue column should remain as well.

Each note that is entered should stay visible on this page, but only during this session. That way the user can edit the note easily.

In a separate frame at the bottom 1/4 of the page (keeping it visible on the page at all times) should be the same Potential and Proposed resources function I've created/proposed in the "Discussion Review 1" issue. It will change/update with each Opportunity line that is clicked/selected. To be clear, this will not require the user clicking the Oppty Name, but rather just clicking any part of the line that isn't linked.