Comments - Defect Number: 2667; Defect RecordID: 2663; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Projects>Add Project: Add ability to upload/attach SOW .pdf document to Project; Defect Description: I think it would be very useful to implement an upload/attach feature to the Add Project function so that a SOW .pdf could be linked to the Project. I think it should be restricted to .pdf files only (so that they're read-only) and only allow 1 file to be attached. If the user attempts to upload a 2nd/newer SOW they should be alerted that one already exists and that if they continue that file will be replaced with this one (buttons on pop up message: "Continue" "Cancel").
We need to consider how this will affect database size in relation to Azure. It should be small files, size-wise, I'd think.

Also, update the color of the icon for each project in this way: If the Project is Active and has a SOW attached, the icon will be fully GREEN (as it is now). If the Project is Active but does not have a SOW attached, the icon will be HALF GREEN, HALF AMBER (split vertically). My plan is to phase out "Projected" as a Project status (projected projects should be considered opportunities is my thinking).