Comments - Defect Number: 2668; Defect RecordID: 2664; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Configuration>Financial>Expense Categories: UI/Layout/Functionality changes; Defect Description: Currently when the user lands on the Expense Category page, if nothing has been created they see this:
Please change this to appear by default as the table that would show if even one category has been created. It should function like other tables recently changed (Practices, for example):

It should have alternating colors for each line, like other tables. Include the pencil icon (1) to denote editing (include hover explanation of icon). When the pencil is selected, change the icons to the "check" and "no" (2). Replace the existing "Add Expense Category" button with the green "+" icon. When clicked, the "+" should be replaced by the green "check" and red "no" icons, and show a box with the "Add Expense Category" watermark that disappears when the user clicks into the text box. When the user clicks the green "check" to save the new category it should just add to the table in the correct alphabetical position (ascending).