Comments - Defect Number: 2764; Defect RecordID: 2760; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Configuration>Company Resources>Opportunity Priorities: Create page/functionality, tie to Opportunity pages; Defect Description: I'd like to add this page so that a company can optionally define what the opportunity priorities mean to them. They'll generally just be a text box description, but depending on what is selected on this page, these priorities should populate in the drop-down when users add new opportunities. The green plus/add button should enable the drop-down (it would be greyed-out otherwise). The drop-down should show all letters that haven't already been used. By default (for new setups) only A, B, C, and D should be enabled with no description entered. Once the green plus/add button is clicked the icon should change to the green check mark and red "no" icon (hover tag for each icon should say "Save" or "Cancel", respectively). Clicking the green plus/add button should also enable the text box in the description column so that a description can be entered if desired. Limit the characters to 100, no validation necessary.
There should not be a restriction on which letters can be used. For example, a company should be able to define A, C, F, Z if that's how they want to organize their priorities. Whatever is chosen, it should always be alphabetically listed ascending. A company only really needs one priority at least.

If an admin clicks the pencil icon to change an existing priority's letter they should be able to. If the priority the admin is changing is already linked to any opportunities in PM, when they click the green check mark to save the change they should get a pop-up message stating "This priority is linked to existing opportunities already. Changing this value will also change the status for all existing opportunities that are using the old value." Clicking the OK button in the pop-up should then apply the change. If the priority being changed is not linked to any opportunities in PM then they should not get a pop-up message.

If an admin clicks the red x icon to delete an existing priority's letter they should be able to as well, but they should get a pop-up message stating "This priority is linked to existing opportunities already and can not be deleted without first assigning an new priority to these opportunities. Please select a replacement from the drop-down below then click OK." As implied, the pop-up message should have a drop-down box that contains all active/valid priorities (except for the one being deleted). Clicking OK will assign this newly selected prioritiy to opportunities and then delete the priority.

On the Add Opportunity page, add a little (?) icon to the right of Priority (before the drop-down selection box) similar to what is on the Project page and used for the Monthly Status fly-out. When the user clicks that icon it should fly-out and list each active priority and the description as entered on the Oppty Priority page. Also include this icon and functionality on the main Oppty landing page next to the Priority header.