Comments - Defect Number: 2704; Defect RecordID: 2700; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Configuration>Company Resources: Create "Opportunity Priorities" page; Defect Description: Need to add an "Opportunity Priorities" configuration page. From that page the Admin should be able to define what A B C D are (text) for the company, or which of those letters are used (via a checkbox, all selected by default). Characters should be somewhat limited - 50 max, perhaps.
On the Opportunity page add a "Legend" bar/drop so that the users can quickly refresh themselves what each priority stands for. This will populate with whatever has been chosen/entered/saved on the config page.

On the Add Opportunity page, when the user uses the drop-down to select the Priority, can it show the definition of each Priority as well, wrapping if necessary (so that we don't have wide drop-down)? EX:

[x]A - Text text text
[x]B - Text text text
    text text text
[x]C - Text Text
[x]D - Text Text

[Save All]