Comments - Defect Number: 2877; Defect RecordID: 2873; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Opportunity>Discussion Review 2: Minor expansion to Proposed Resources selection box; Defect Description: I'd like to make the current selection checkbox the "check" column as in the image below. This functions exactly the way it does currently and makes the person name appear in the DR2 page. If this is checked, the X checkbox cannot be checked.
The X column would make the name appear with a strikethrough on the DR2 page. If this is checked, the check checkbox cannot be checked.

These checkboxes also need to appear in the Add Oppty and Oppty Detail pages. When converting an Oppty to a Project, only the names with the "check" checkbox selected should carry over.


Alternatively, is it possible or easier to do this with only a single column of checkboxes, where the first click in the box selects the name, a second click would strikethrough the name, and a third click would clear the name entirely? The progress would look like this:

Name[ ] (default, not selected)
Name[ / ] (one-click, selected)
Name[ X ] (two clicks, strikethrough)
Name[ ] (three clicks, back to default)