Comments - Defect Number: 2876; Defect RecordID: 2872; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Reports>Resources>Consultant Util Timeline: Change to week divisions; Defect Description: We need to update the parameters bar to as in the image below (similar to Project Summary parameter bar change for date range).
The first drop down should have:
* Next 3 Months (default)
* Last 3 Months
* Custom Dates
*When selected prompt user to select From and To date via calendar

The "by" dropdown should just be:
* Day
* Week
* Month

Next we need to tweak the week divisions. Currently if the current month is June, the calendar will start on June 1, regardless of what day that falls on, then the divisions are +7 days at a time. What we need to do instead is reflect the actual Sunday to Saturday week we have defined (same as in Time Entry).

If we ran the report today (in June) with default settings (Next 3 Months, by WEEK), the graph would actually start on 5/30 and go to 9/3. If viewing by WEEK, the individual weeks would still be shown with the weekly utilization calculated and color coded correctly (Ex: for that first week, it would be 5/30-6/4), but for the overall U% it would be calculated from 6/1-8/31.

Another example, if this is run in July, it will by default (Next 3 Months, by WEEK) show 6/27-10/1. (Note: these are Sunday to Saturday weeks, but of course the weekend days do not negatively affect U%). The first week utilization will be calculated for 6/27 to 7/2 (since July 1 falls on a Friday) to show the appropriate color coding, but overall U% will still only reflect 7/1-9/30.

If a Custom Date is chosen, the TO calendar should have days grey/ineligible for selection that are beyond 3 months out from the FROM date selected. However, it should be possible for someone to select a short range, for example, from 7/1-7/15. If viewing by WEEK, the graph should still be from 6/27 to 7/16 (showing entire weeks). Overall U% should only reflect 7/1-7/15. If viewing by DAY, the graph should only show 7/1 to 7/15. If someone selects by MONTH, just automatically change the range to 7/1-7/31, but display the graph correctly (6/27-8/6). Overall U% should only reflect 7/1-7/31.