Comments - Defect Number: 2882; Defect RecordID: 2878; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Configuration>Company Resources>Clients: Create "Margin Goals" tab/section for each Client, color code margin thresholds; Defect Description: We need a section created for each CLIENT so that either the CLIENT DIRECTOR or ADMINISTRATOR role can define Margin goals (by percentage) for each CLIENT. This helps to indicate to Managers, etc, that are prohibited from seeing actual Margin totals that their project is acceptably profitable. It can be a tab next to "Filters" in each Client detail page.
It should be possible to toggle this color-coding feature off/on (UNCHECKED/OFF by default). While toggled off, the section should appear as it does in the image below (above the blue dash line), but be greyed out.

When toggled on, the admin should be able to select percentages from the drop-downs, as well as assign a color to that percentage.

Limit the number of thresholds to 5 max, so only 5 distinct colors are needed in the color selection.

The percentages selected cannot overlap. The same color cannot be chosen. If it is possible to fade out a previously selected color from the selection drop-down to indicate visually that it cannot be selected again, do that.

It should be possible to delete a threshold and adjust the remaining thresholds.

If thresholds are toggled on, they must always add up to 100%, but it should be possible to go as high as 150%.

Secondly, these newly added threshold colors should appear in the following areas:
1. Surrounding/highlighting the Milestone financial summary Margin percentage

2. Highlighting the Project financial summary Margin percentage (also, remove the word "Margin" and put in parentheses, as it appears in the Milestone financial summary)

3a. Highlighting the Margin percentage on the Margin Test page:

3b. Currently the persons that any single user can see on the Margin Test page is governed by their security role setting. We need to change this so that all users with rights to the Margin Test page see active/projected persons, but depending on their security role they cannot see the actual salary and calculated values. They can change the bill rate and hours still, but calculated values should read as "(hidden)" as they do throughout PM, but the user could see the color coded highlighting.