Comments - Defect Number: 2803; Defect RecordID: 2799; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Projects: Create Export functionality; Defect Description: Create a function to Export the filtered Projects list to Excel for the following fields:
ProjectIDClientClient-GroupBuyerProjectNameStatusPracticeRevOwnerSalespersonDirector
Mar

Make sure the exported spreadsheet includes the date range at the top as well. Any financial details that are (hidden) for the user should also be (hidden) in the export as well.

If it's possible to have Revenue and Margin in the same box (the way it appears in Projects already) that'd be ideal.