Comments - Defect Number: 2829; Defect RecordID: 2825; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Projects: Add Calendar view, allow all roles to view; Defect Description: This will appear very similar to the detail a user can see for each Person on the Consultant Util Timeline report.
It will just be a simple graph, showing a bar for each active and projected project over a 6 month period. The intention is to just allow everyone to get an idea of what is happening with projects, so all roles should be able to see this. We may expand the functionality later, but for now it's view-only (nothing linked). Only show Client name and Project name on the bar, unless the bar is short (project duration is short) in which case just show the P#. Keep the color coding the same as it is on the Summary page (Green = Active, etc).