Comments - Defect Number: 2862; Defect RecordID: 2858; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: TimeEntryReport.aspx:  User with role Practice Area Manager is able to open the ProjectDetails and MilestoneDetails page of a @; Defect Description: Steps to Repro:
1) Login as Practice Area Manager.
2) Open TimeEntryReport.aspx page.
3) Select the dates which has the time entries.
4) Click on Update button.
  Result: GTE Report will be displayed.
5) Click on any ProjectDetails/Milestone details page of a project linked to the client on which user dont have permissions

Result: User is navigated to ProjectDetails/MilestoneDetails page of a project linked to the client on which user dont have permissions.
Expected: User with Practice Area Manager Role should not have Access to the ProjectDetails and Milestone details page of a project linked to the client on which user dont have permissions.

Note: Same issue exists with Client Director Role.