Comments - Defect Number: 2716; Defect RecordID: 2712; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Configuration>Security: Add "General" page; add "Lockout Policy" and "Password Policy" sections; Defect Description: Under the Security subheading, create a "General" page. On that page an Admin should be able to define the "Password Policy" that affects all users, and a "Lockout Policy" that affects all users.
The "Password Policy" should default to a secure password, at least 7 characters, mix of upper/lowercase, numbers and/or special characters. Give the admin the option to choose from a drop-down how many previous passwords are remembered and can't be used again (ex, cannot reuse any of the last 3 passwords by default) and how often the password can be changed (no limit, not more than once in 24 hours, not more than once in 1 week).

The "Lockout Policy" should default to if incorrect password entered [3] time(s) in [15 minutes], lock user account. (Values between brackets would be drop-down options for Admin). Also need a checkbox to either enable or disable policy entirely (checked/enabled by default), and another checkbox to notify user or not when account is locked (checked/enabled by default). The template e-mail should be added to the E-mail Templates page as well ("Lockout Notification") under System Templates. It should include a link to the reset password page/functionality (built in #2700). It can be defined in the template if an Admin wants to add their address (or a helpdesk address) to the CC: line.

I'd like the "General" page layout to look like the one I mocked up in #2645 so that the functionality can be explained on the right-side of the page, and the options can be defined on the left side.

If any content needs to be defined I will write and add.