Comments - Defect Number: 2760; Defect RecordID: 2756; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Configuration>Time>Notes:  Create page, add functionality; Defect Description: I'd like to create a page at Configuration>Time Entry>Notes where and Admin can define whether notes are a requirement in Time Entry. By default the "Notes are required" box should be checked and all active practices moved to the "Must Enter Notes" column. When it is checked, the greyed out Practice names listed in the "Exempt from Notes" selection box will be normal text colored (black) and will be individually selectable. They will also automatically all be added to the "Must Enter Notes" column. The arrows between the boxes signify "Add All", "Add Selected", "Remove Selected", and "Remove All". These labels should be shown when the user hovers over the buttons.
The two boxes where active practices are listed should be set at a reasonable height and scroll if necessary. If we have a character limit for the Practice name defined elsewhere, these boxes should be at least as wide as the character width so that there is no horizontal scrolling.

If notes are required but a particular practice is left in the "Exempt from Notes" column then any users linked to that practice should not be required to enter any notes for their time. They still can, optionally, so no icons should be removed for any users on the Time Entry page.

If notes are required and the practice is in the "Must Enter Notes" column then any users linked to that practice should be required to enter notes for their time. There should also be a message on the Time Entry page for those users that clearly states "You are required to enter notes for any/all time you enter on this page."

No matter what boxes are checked/unchecked or what practices are moved, no changes are committed until the Admin/user clicks the "Save Changes" button. It should be greyed out if there are no changes made to the page. Once clicked, there should be a "Changes saved successfully" message shown to the Admin/user.

**4/20/11 UPDATE** If a new practice is added, it should, by default, be added to the "Must Enter Notes" column.