Comments - Defect Number: 2890; Defect RecordID: 2886; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Projects>Milestones>Persons: Owner need to be able to adjust personhours allocation regardless of seniority; Defect Description: This is similar to issue #2588.
Anyone defined as the Owner of a Project needs to have the ability to change hour allocation for Resources/Persons assigned to the Milestones, regardless of the seniority difference between the Owner and the Resources/Persons. They don't need to see the Margins still, but they should be able to reallocate time. They should still not be able to completely remove the Resource/Person if the Person has already entered time/notes towards the Milestone.