Comments - Defect Number: 2899; Defect RecordID: 2895; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Reports>Resources: Bench Costs reporting different values for single month when run by itself vs in a range of months; Defect Description: Please see below.
The three names that are not on the August-only report, but appear with costs for August on the August-to-September report, are Usha Anumolu, Shawn Gahr, and Shivika Garg.

I ran the bench costs report in PM and its showing a different total for August when I run it alone vs. running alongside September. Can you please look into why they are different? Thanks.