Comments - Defect Number: 2894; Defect RecordID: 2890; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Time Entry>Calendars: Days selected on person's calendar as PTO should auto-enter in Time Entry; Defect Description: This is very similar to the Company Holiday changes.
When a day is selected on an individual person's calendar to note the day as PTO, that day should be automatically pre-entered in a row on the person's calendar, with the note pre-populated as PTO. The other days in this row that are not PTO should be dark grey/ineligible for time/note entry.

We should also remove the PTO Time Type from the drop down list as the user will not need to select it.