Comments - Defect Number: 2733; Defect RecordID: 2729; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Reports>Time Entry>Time Entry by Person: UI issues; Defect Description: Preconditions: create a person "test user" and save Time entries for current week. Make sure that person have vacation, company holiday, but the person is scheduled to work anyway for current week(view Legend for more details).
Steps to Repro:
1) Open Time Entry->Time Entry by Person report.
2) Click on "Update View" button to view the details of person "test user".
3) View the details.

a) IE, Firefox(3.6.12) ,chrome(9.0.587.0): View the background color for holidays\vacation days.
Result: Background highlighting color is alligned to left.
Expected: Background color for any holiday should be alligned center.

b) IE, Firefox(3.6.12) ,chrome(9.0.587.0): View the background color of vacation days, company holidays but person scheduled to work on that day.(This is not specified in screen shot)
Result: Background color for vacation is not displayed(white) and company holiday, but the person is scheduled to work anyway is displayed in pink.
Expected: Background color for vacation should be displayed in orange and company holiday, but the person is scheduled to work anyway should be displayed in green.

C) Firefox(3.6.12) ,chrome(9.0.587.0): View the TimeEntry Summary Table.
Result: Time Entry Summary table width is not stretched to right of the page.
Expected: Time Entry Summary table width should be stretched to right of the page.

Please find the screenshot below for more details.