Comments - Defect Number: 2759; Defect RecordID: 2755; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Configuration>Troubleshooting>Activity Log: Move existing functionality/page here; Defect Description: Move the existing Activity Log page to Configuration>Troubleshooting>Activity Log and replace the current parameters bar with the following:
There are no filters to be used, so no need for a chevron filter expansion. Replace the current "over period" drop-down with "From" and "To" date selection boxes. By default the From date should populate with the current date minus 1 year, and the To box should be the current date. Grey out the "Reset Filter" button if only defaults are selected.