Comments - Defect Number: 2912; Defect RecordID: 2908; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Reports>Time>TE by Person/Project: Create export to Excel/PDF functionality; Defect Description: Based on request in e-mail. Just need ability to export both reports to Excel and PDF, preferably formatted with appropriate page breaks.
For the TE by Project export, we need to include the "Client - Project Name - Project Number - Milestone - Date Range" on each page. Each employee would be separated by a page break.

For the TE by Person report (currently working in staging) we need to eliminate the empty columns between the summary days.

If it is possible to define export orientation and margin formatting, I believe landscape would work best, with narrow margins. For TE by Person, a 15-day range is normal, and it would be ideal if all columns fit on a single page.

Finally, opening any Excel file that PM generates results in the following message:


Is there a way to resolve this?