Comments - Defect Number: 2913; Defect RecordID: 2909; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Create role-based dashboard landing pages; Defect Description: Initially we can identify seven role-based dashboards: 1. Consulant Dashboard (applies to Intern, Analyst, Consultant, Sr. Consultant) 2. Manager Dashboard (applies to Manager, Sr. Manager) 3. Business Development Dashboard (applies to Business Developement Manager) 4. Client Director Dashboard (applies to Client Director) 5. Senior Dashboard (applies to Principal, President) 6. Recruiter Dashboard (applies to Recruiter) 7. Admin Dashboard (applies to Admin)
For now, the dashboard landing page for each should include the Practice Management Announcements section and a Quick Liinks section. Beneath that will be content that is specific to each role-based dashboard. We are still determining what information is best for each, but the idea will be to keep it somewhat high-level and simplified to keep loading times down.

In the Practice Management Annoucements section, this should be a section that an Admin can edit, but will be shown on all dashboards. Is it possible to have an "Edit" button that applies only to an Admin view, along with a "Save" button (maybe this is done in a modal dialog)? When "Edit" is clicked can we have normal editing controls (underline, bold, italics, hyperlink, font size, etc)?

It would also be good for an Admin user to have a dropdown on the page so that they can switch between dashboard pages in order to edit content specific to each, such as the Quick Links. We would need to be able to edit and add/remove links to this section.

Quick Links for each role are as follows:
1. Time Entry; Request Time Off (e-mail link to person's Line Manager with subject "Time Off Request")
2. Add Project; Time Entry by Project; Margin Test
3. Add Opportunity; Discussion Review; Add Project; Margin Test
4. Add Opportunity; Discussion Review; Add Project; Group by...; Margin Test
5. Group by...; Discussion Review; Consultant Util Timeline; Margin Test
6. Discussion Review; Persons; Margin Test
7. Persons; Activity Log; Become User

Is it also possible to include differerent search boxes in a section above Quick Links? Or, alternatively, one search box with a drop down next to it where the type of search can be chosen (Project/Opportunity/Person, etc) These would appear on only certain dashboards, and be relevant to the role:

1. No search boxes.
2. Project search box
3. Opportunity, Project search box
4. Opportunity, Project search box
5. Opportunity, Project search box
6. Persons search box
7. Opportunity, Project, Persons search box