Comments - Defect Number: 2920; Defect RecordID: 2916; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Scrolling Drop down UI issues throughout PM.; Defect Description: Steps to repro:
1) Open Persons.aspx page.
2) Click on Expand filters chevron and go to filters tab.
3) Scroll down a bit.
4) Now expand owners drop down.
Result: Page will be scrolled to the top and drop down is not expanded.
Expected: Page should not be scrolled to the top and drop down should be expanded.

Steps to repro:

1) Open Persons.aspx page.
2) Click on Expand filters chevron and go to filters tab.
3) Minimize the page by clicking on restore a bit.
Result: Horizontal scroll bar is displayed for the page.
4) Scroll to right by using Horizontal scroll bar.
5) Now expand owners drop down.
Result: Drop down is displayed to the right. (See image below)
Expected: Drop down should be displayed at the correct place.