Comments - Defect Number: 2758; Defect RecordID: 2754; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Configuration>Troubleshooting>Become User:  Move existing functionality to new page, slight UI change; Defect Description: I'd like to move the existing Become User functionality to it's own page under the Troubleshooting subcategory.
Is it possible that when the Admin selects the user to become and clicks OK that a new window is opened for PM with that user logged in? That way the Admin can simply close that new window and go back to using PM as themselves without having to Logout and then log back in.