Comments - Defect Number: 2930; Defect RecordID: 2926; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Reports>Time>Time Entry by Project: Change to filter logic; Defect Description: FIRST PART: Currently several areas in PM use filter logic that depends on what "Project/Opportunity Permissions" are selected per person in Config Resources>Resources>Person. This is flawed because it is not efficient and does not allow us to give persons access to specific, relevant projects without giving them access to other projects under the same grouping.
We can begin with Time Entry by Project to change this. For persons with Admin or Senior Leadership roles the page should continue to show all data it currently does.

For all other roles, the page should determine what clients and projects are relevant based on the relationship of the person that is logged in to projects, based on a few key fields in the project.

The key fields are Client Director, Salesperson, and Project Manager(s) (previously called "Owner"). If the logged in user is assigned to any of those roles on a project, then the client dropdown in the Time Entry by Project page should have the Client for that Project available.

The report output should appear as it does now, and allow the user to see all time and notes entered by the project/milestone resources.

SECOND PART:
Add a checkbox beneath the Project dropdown selector for "Active and Internal Projects only" that is checked by default. When unchecked this would allow the user to see Completed projects as well. Here's a reference: