Comments - Defect Number: 2941; Defect RecordID: 2937; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: PM Config: Project Lead role should have access to Project Summary page; should see Projects they are Owner of; Defect Description: We have the ability to specify now that a person with the Project Lead role is a "Project Manager" for a project, but because that role does not have access to the Project Summary page, they are unable to access the Project.
We need to give persons with the Project Lead role access to the Project Summary page (as well as Project Search), but limit their view to be only projects that are relevant to their person, ie ones they are assigned as Project Manager.

They should not have the ability to Add Project, and do not need to see Opportunities. They should only see reports for Time Entry by Project and Time Entry by Person.

We may need to add an additional dashboard view as well, so that we can remove the "Add Project" link from their dashboard.