Comments - Defect Number: 2952; Defect RecordID: 2948; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Time>Calendar: User should see themselves and employees where they are the designated "Line Manager"; rename "Line Manager" to@; Defect Description: Currently a logged in user sees their own name in the calendar dropdown, followed by names of employees that are junior to them.
We need to change this so that the dropdown lists contents are relevant to the logged in user. It should display the logged in user as well as any employees they are linked to as the designated "Career Counselor".

This should have a hierarchical effect, example:

Person A is Career Counselor for Person B, Person C, Person D

Person M is Career Counselor for Person E, Person F, Person G

Person Z is Career Counselor for Person A, Person M

When Person Z logs into PM and is on the Calendar page, they should see their name as well as those of Person A, Person M, and all of the persons for Person A (B, C, D) and Person M (E, F, G) listed.

Persons with SLT, HR, or Admin roles should see all names.