Comments - Defect Number: 2925; Defect RecordID: 2921; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Time Entry: Duplication of hours due to resources issue on project; Defect Description: Original e-mail is below. I believe this issue is due to an issue in the Resources tab. PM will not allow a resource to be removed from a milestone if the person has already entered time and notes towards the milestone, which has always been the case, but the issue occurs when a person is added to a milestone more than once for overlapping periods. It is not possible to remove the person from a milestone, even when it is a redundant entry.
For example:
The current date is 9/27
MilestoneA = 9/1-9/30
PersonA is added as a resource to MilestoneA for 9/1-9/30
PersonA is also added as a resource to MilestoneA for 9/20-9/30
PersonA has entered time and notes for MilestoneA from 9/1-9/21
It is now not possilble to remove the second entry for PersonA from MilestoneA because PM sees time and notes entered through 9/21 and does not allow the removal.

I checked the time report (copied below) which says Kerissa has 121hrs booked but every entry is repeated, so her hours are in fact 60.5 which matches my expectations and what I believe we should bill. The report you (Ivana) have got, gives over 70hrs and contains duplicate entries for 19-21st (totaling 14hrs).