Comments - Defect Number: 2961; Defect RecordID: 2957; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Time > Calendar : Change permissions; Defect Description: Change person dropdown on Calendar page so that all users have read-only permissions to all calendars, but retain the current functionality that gives Career Counselors read-write permissions to the persons linked to them.
Add a note to calendars when they are being displayed as read-only: "ALERT: You are viewing this calendar as READ-ONLY. If you believe you should have permissions to make changes to this calendar, please [contact support]." [contact support] should be a mailto: link point to the support e-mail address with a subject of "Permissions for [Person name]'s calendar".

If the person has edit (read/write) permissions to the particular calendar, do not show the ALERT message.

EX: