Comments - Defect Number: 2969; Defect RecordID: 2965; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Reports->Time->Time Entry by person: Exported Excel format needs to be formatted. See Attached excel sheet for the new format.; Defect Description: The format of the Exported Excel report should be formatted same as the format in the attached Excel sheet.
Email sent by Jason for this issue:

From: Jason Hong-Turney [mailto:jasonht@logic2020.com]
Sent: Tuesday, December 06, 2011 6:08 AM
To: MadhanA (madhana@ggktech.com)
Subject: Timeout issues in TE by Person report

Hi Madhan,

Im trying to run a report for all W2-Salary employees, not in the Admin practice, from 1/1/11 to 11/30/11, but the report times out (also, unselecting filters results in undesired names still appearing both in the person dropdown list as well as the report output). I did manage to get the report to run once, but was completely unable to export the results to Excel.

Also, it would be ideal for the Excel output to be not formatted like a report (leave that to PDF), but instead formatted like a true table.

As soon as we can have this working, please update PM.

Jason Hong-Turney // Logic20/20 // jasonht@logic2020.com
t.206-576-0472 // c. 425-390-4664