Comments - Defect Number: 2959; Defect RecordID: 2955; Server Address: services.logic2020.com; Server Port: 99; Database Name: Practice Management; Defect Summary: Time Entry / Calendar : PTO should have a bi-directional relationship; Defect Description: When employees enter PTO directly into Time Entry, it needs to be noted/highlighted on their Calendar page.
When the relationship exists, clicking the PTO day from the calendar should bring up the dialog box as an edit function, so that a single click doesn't remove the entry from the calendar. This way a manager could simply change 8 hours to 4 hours, for example. We may need to add a button to the dialog box to "Delete" the entry. ("OK" "DELETE" "CANCEL", where Delete can be inactive if there isn't anything to delete)

There should be a hover for any PTO days on the calendar that indicates how many hours are entered as well, for example: "PTO - 4 Hrs" or "PTO - 2.5 Hrs"

There should never be more than one PTO row on a person's Time Entry page.